add day 19
The second par of day 19 was calculated by looking at solution megathread.

The input calculates a big number and sums its all factors. For this it uses the following method:

int SUM = 0

for(int i=0; i<NUM; ++i) {
    for(int j=0; j<NUM; ++j) {
        if(i*j == NUM) {
            SUM += i
        }
    }
}

it is very slow and inefficient (even == and * is done as slow as many steps as possible) the good solution for it would be to write a code that sees this and optimizes it but it is not the case